refactor: split openrouter account flow and detail tokens
